refactor(wshandler): move window_watcher into the wshandler package (#2858)
